mm-kasan-switch-slub-to-stackdepot-enable-memory-quarantine-for-slub-fix
make calculate_sizes():size a size_t

Reported-by: kbuild test robot <lkp@intel.com>
Cc: Alexander Potapenko <glider@google.com>

+++ mm/slub.c
@@ -3387,7 +3387,7 @@ static void set_min_partial(struct kmem_cache *s, unsigned long min)
 static int calculate_sizes(struct kmem_cache *s, int forced_order)
 {
 	unsigned long flags = s->flags;
-	unsigned long size = s->object_size;
+	size_t size = s->object_size;
 	int order;
 
 	/*
